Rollback changes made on gtest

+++ tests/gtest/include/gtest/internal/gtest-port.h
@@ -198,7 +198,6 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include <stdint.h>
 #ifndef _WIN32_WCE
 # include <sys/types.h>
 # include <sys/stat.h>
@@ -1647,7 +1646,7 @@ typedef __int64 BiggestInt;
 #else
 # define GTEST_PATH_SEP_ "/"
 # define GTEST_HAS_ALT_PATH_SEP_ 0
-typedef int64_t BiggestInt;  // NOLINT
+typedef long long BiggestInt;  // NOLINT
 #endif  // GTEST_OS_WINDOWS
 
 // Utilities for char.
@@ -1892,8 +1891,8 @@ class TypeWithSize<8> {
   typedef __int64 Int;
   typedef unsigned __int64 UInt;
 #else
-  typedef int64_t Int;  // NOLINT
-  typedef uint64_t UInt;  // NOLINT
+  typedef long long Int;  // NOLINT
+  typedef unsigned long long UInt;  // NOLINT
 #endif  // GTEST_OS_WINDOWS
 };
